backend/requirements.txt
Flask==2.2.5
flask_sqlalchemy==3.0.3
python-dotenv==1.0.0
requests==2.31.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
flask-jwt-extended==4.5.2
transformers==4.34.0
torch==2.2.0
librosa==0.10.0
soundfile==0.12.1
numpy==1.26.0
python-multipart==0.0.6
pydub==0.25.1
